+++ apps/dash-object-detection/requirements.txt
@@ -1,9 +1,9 @@
 dash==1.0.0
 gunicorn==19.9.0
 pillow==8.2.0
-scipy==1.2.1
-numpy==1.16.1
-pandas==0.24.1
-Flask==1.0.1
+scipy==1.8.0
+numpy==1.22.3
+pandas==1.4.2
+Flask==2.1.2
 dash-player==0.0.1
 pathlib==1.0.1
