+++ apps/dash-opioid-epidemic/constants.py
@@ -1,0 +1,71 @@
+import os
+import pathlib
+import pandas as pd
+
+
+# Load data
+
+APP_PATH = str(pathlib.Path(__file__).parent.resolve())
+
+df_lat_lon = pd.read_csv(
+    os.path.join(APP_PATH, os.path.join("data", "lat_lon_counties.csv"))
+)
+df_lat_lon["FIPS "] = df_lat_lon["FIPS "].apply(lambda x: str(x).zfill(5))
+
+df_full_data = pd.read_csv(
+    os.path.join(
+        APP_PATH, os.path.join("data", "age_adjusted_death_rate_no_quotes.csv")
+    )
+)
+df_full_data["County Code"] = df_full_data["County Code"].apply(
+    lambda x: str(x).zfill(5)
+)
+df_full_data["County"] = (
+    df_full_data["Unnamed: 0"] + ", " + df_full_data.County.map(str)
+)
+
+
+YEARS = [2003, 2004, 2005, 2006, 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015]
+
+BINS = [
+    "0-2",
+    "2.1-4",
+    "4.1-6",
+    "6.1-8",
+    "8.1-10",
+    "10.1-12",
+    "12.1-14",
+    "14.1-16",
+    "16.1-18",
+    "18.1-20",
+    "20.1-22",
+    "22.1-24",
+    "24.1-26",
+    "26.1-28",
+    "28.1-30",
+    ">30",
+]
+
+DEFAULT_COLORSCALE = [
+    "#f2fffb",
+    "#bbffeb",
+    "#98ffe0",
+    "#79ffd6",
+    "#6df0c8",
+    "#69e7c0",
+    "#59dab2",
+    "#45d0a5",
+    "#31c194",
+    "#2bb489",
+    "#25a27b",
+    "#1e906d",
+    "#188463",
+    "#157658",
+    "#11684d",
+    "#10523e",
+]
+
+DEFAULT_OPACITY = 0.8
+
+mapbox_access_token = "pk.eyJ1IjoicGxvdGx5bWFwYm94IiwiYSI6ImNrOWJqb2F4djBnMjEzbG50amg0dnJieG4ifQ.Zme1-Uzoi75IaFbieBDl3A"
+mapbox_style = "mapbox://styles/plotlymapbox/cjvprkf3t1kns1cqjxuxmwixz"
