+++ apps/dash-opioid-epidemic/requirements.txt
@@ -1,6 +1,5 @@
-dash==1.12.0
-plotly==4.7.1
+dash==2.4.1
+pandas==1.4.2
+gunicorn==20.1.0
+plotly==5.8.0
 cufflinks==0.17.3
-gunicorn==20.0.4
-numpy==1.18.4
-pandas==1.0.3
