+++ apps/dash-opioid-epidemic/utils/helper_functions.py
@@ -1,0 +1,220 @@
+import re
+import pandas as pd
+
+from constants import (
+    df_full_data,
+    df_lat_lon,
+    BINS,
+    DEFAULT_COLORSCALE,
+    DEFAULT_OPACITY,
+    mapbox_access_token,
+    mapbox_style,
+)
+
+
+def display_map(year, figure):
+    cm = dict(zip(BINS, DEFAULT_COLORSCALE))
+
+    data = [
+        dict(
+            lat=df_lat_lon["Latitude "],
+            lon=df_lat_lon["Longitude"],
+            text=df_lat_lon["Hover"],
+            type="scattermapbox",
+            hoverinfo="text",
+            marker=dict(size=5, color="white", opacity=0),
+        )
+    ]
+
+    annotations = [
+        dict(
+            showarrow=False,
+            align="right",
+            text="<b>Age-adjusted death rate<br>per county per year</b>",
+            font=dict(color="#2cfec1"),
+            bgcolor="#1f2630",
+            x=0.95,
+            y=0.95,
+        )
+    ]
+
+    for i, bin in enumerate(reversed(BINS)):
+        color = cm[bin]
+        annotations.append(
+            dict(
+                arrowcolor=color,
+                text=bin,
+                x=0.95,
+                y=0.85 - (i / 20),
+                ax=-60,
+                ay=0,
+                arrowwidth=5,
+                arrowhead=0,
+                bgcolor="#1f2630",
+                font=dict(color="#2cfec1"),
+            )
+        )
+
+    if "layout" in figure:
+        lat = figure["layout"]["mapbox"]["center"]["lat"]
+        lon = figure["layout"]["mapbox"]["center"]["lon"]
+        zoom = figure["layout"]["mapbox"]["zoom"]
+    else:
+        lat = 38.72490
+        lon = -95.61446
+        zoom = 3.5
+
+    layout = dict(
+        mapbox=dict(
+            layers=[],
+            accesstoken=mapbox_access_token,
+            style=mapbox_style,
+            center=dict(lat=lat, lon=lon),
+            zoom=zoom,
+        ),
+        hovermode="closest",
+        margin=dict(r=0, l=0, t=0, b=0),
+        annotations=annotations,
+        dragmode="lasso",
+    )
+
+    base_url = "https://raw.githubusercontent.com/jackparmer/mapbox-counties/master/"
+    for bin in BINS:
+        geo_layer = dict(
+            sourcetype="geojson",
+            source=base_url + str(year) + "/" + bin + ".geojson",
+            type="fill",
+            color=cm[bin],
+            opacity=DEFAULT_OPACITY,
+            # CHANGE THIS
+            fill=dict(outlinecolor="#afafaf"),
+        )
+        layout["mapbox"]["layers"].append(geo_layer)
+
+    fig = dict(data=data, layout=layout)
+    return fig
+
+
+def display_selected_data(selectedData, chart_dropdown, year):
+    if selectedData is None:
+        return dict(
+            data=[dict(x=0, y=0)],
+            layout=dict(
+                title="Click-drag on the map to select counties",
+                paper_bgcolor="#1f2630",
+                plot_bgcolor="#1f2630",
+                font=dict(color="#2cfec1"),
+                margin=dict(t=75, r=50, b=100, l=75),
+            ),
+        )
+    pts = selectedData["points"]
+    fips = [str(pt["text"].split("<br>")[-1]) for pt in pts]
+    for i in range(len(fips)):
+        if len(fips[i]) == 4:
+            fips[i] = "0" + fips[i]
+    dff = df_full_data[df_full_data["County Code"].isin(fips)]
+    dff = dff.sort_values("Year")
+
+    regex_pat = re.compile(r"Unreliable", flags=re.IGNORECASE)
+    dff["Age Adjusted Rate"] = dff["Age Adjusted Rate"].replace(regex_pat, 0)
+
+    if chart_dropdown != "death_rate_all_time":
+        title = "Absolute deaths per county, <b>1999-2016</b>"
+        AGGREGATE_BY = "Deaths"
+        if "show_absolute_deaths_single_year" == chart_dropdown:
+            dff = dff[dff.Year == year]
+            title = "Absolute deaths per county, <b>{0}</b>".format(year)
+        elif "show_death_rate_single_year" == chart_dropdown:
+            dff = dff[dff.Year == year]
+            title = "Age-adjusted death rate per county, <b>{0}</b>".format(year)
+            AGGREGATE_BY = "Age Adjusted Rate"
+
+        dff[AGGREGATE_BY] = pd.to_numeric(dff[AGGREGATE_BY], errors="coerce")
+        deaths_or_rate_by_fips = dff.groupby("County")[AGGREGATE_BY].sum()
+        deaths_or_rate_by_fips = deaths_or_rate_by_fips.sort_values()
+        # Only look at non-zero rows:
+        deaths_or_rate_by_fips = deaths_or_rate_by_fips[deaths_or_rate_by_fips > 0]
+        fig = deaths_or_rate_by_fips.iplot(
+            kind="bar", y=AGGREGATE_BY, title=title, asFigure=True
+        )
+
+        fig_layout = fig["layout"]
+        fig_data = fig["data"]
+
+        fig_data[0]["text"] = deaths_or_rate_by_fips.values.tolist()
+        fig_data[0]["marker"]["color"] = "#2cfec1"
+        fig_data[0]["marker"]["opacity"] = 1
+        fig_data[0]["marker"]["line"]["width"] = 0
+        fig_data[0]["textposition"] = "outside"
+        fig_layout["paper_bgcolor"] = "#1f2630"
+        fig_layout["plot_bgcolor"] = "#1f2630"
+        fig_layout["font"]["color"] = "#2cfec1"
+        fig_layout["title"]["font"]["color"] = "#2cfec1"
+        fig_layout["xaxis"]["tickfont"]["color"] = "#2cfec1"
+        fig_layout["yaxis"]["tickfont"]["color"] = "#2cfec1"
+        fig_layout["xaxis"]["gridcolor"] = "#5b5b5b"
+        fig_layout["yaxis"]["gridcolor"] = "#5b5b5b"
+        fig_layout["margin"]["t"] = 75
+        fig_layout["margin"]["r"] = 50
+        fig_layout["margin"]["b"] = 100
+        fig_layout["margin"]["l"] = 50
+
+        return fig
+
+    fig = dff.iplot(
+        kind="area",
+        x="Year",
+        y="Age Adjusted Rate",
+        text="County",
+        categories="County",
+        colors=[
+            "#1b9e77",
+            "#d95f02",
+            "#7570b3",
+            "#e7298a",
+            "#66a61e",
+            "#e6ab02",
+            "#a6761d",
+            "#666666",
+            "#1b9e77",
+        ],
+        vline=[year],
+        asFigure=True,
+    )
+
+    for i, trace in enumerate(fig["data"]):
+        trace["mode"] = "lines+markers"
+        trace["marker"]["size"] = 4
+        trace["marker"]["line"]["width"] = 1
+        trace["type"] = "scatter"
+        for prop in trace:
+            fig["data"][i][prop] = trace[prop]
+
+    # Only show first 500 lines
+    fig["data"] = fig["data"][0:500]
+
+    fig_layout = fig["layout"]
+
+    # See plot.ly/python/reference
+    fig_layout["yaxis"]["title"] = "Age-adjusted death rate per county per year"
+    fig_layout["xaxis"]["title"] = ""
+    fig_layout["yaxis"]["fixedrange"] = True
+    fig_layout["xaxis"]["fixedrange"] = False
+    fig_layout["hovermode"] = "closest"
+    fig_layout["title"] = "<b>{0}</b> counties selected".format(len(fips))
+    fig_layout["legend"] = dict(orientation="v")
+    fig_layout["autosize"] = True
+    fig_layout["paper_bgcolor"] = "#1f2630"
+    fig_layout["plot_bgcolor"] = "#1f2630"
+    fig_layout["font"]["color"] = "#2cfec1"
+    fig_layout["xaxis"]["tickfont"]["color"] = "#2cfec1"
+    fig_layout["yaxis"]["tickfont"]["color"] = "#2cfec1"
+    fig_layout["xaxis"]["gridcolor"] = "#5b5b5b"
+    fig_layout["yaxis"]["gridcolor"] = "#5b5b5b"
+
+    if len(fips) > 500:
+        fig["layout"][
+            "title"
+        ] = "Age-adjusted death rate per county per year <br>(only 1st 500 shown)"
+
+    return fig
